// src/types/enums.ts
export enum NodeKind {
  FILE = 'FILE',
  DIR = 'DIR',
  SYMLINK = 'SYMLINK',
  SPECIAL = 'SPECIAL'
}

export enum HashStatus {
  PRESENT = 'PRESENT',
  MISSING = 'MISSING',
  ERROR = 'ERROR'
}

export enum ErrorStage {
  LIST = 'LIST',
  STAT = 'STAT',
  OPEN = 'OPEN',
  READ = 'READ',
  ARCHIVE_LIST = 'ARCHIVE_LIST',
  ARCHIVE_STAT = 'ARCHIVE_STAT',
  ARCHIVE_OPEN = 'ARCHIVE_OPEN',
  ARCHIVE_READ = 'ARCHIVE_READ',
  STORE_READ = 'STORE_READ',
  STORE_WRITE = 'STORE_WRITE',
  EXECUTE = 'EXECUTE'
}

export enum ErrorCode {
  PERMISSION_DENIED = 'PERMISSION_DENIED',
  NOT_FOUND = 'NOT_FOUND',
  IO_ERROR = 'IO_ERROR',
  PATH_TOO_LONG = 'PATH_TOO_LONG',
  BROKEN_SYMLINK = 'BROKEN_SYMLINK',
  ARCHIVE_CORRUPT = 'ARCHIVE_CORRUPT',
  ARCHIVE_ENCRYPTED = 'ARCHIVE_ENCRYPTED',
  ARCHIVE_UNSUPPORTED = 'ARCHIVE_UNSUPPORTED',
  ENCODING_ERROR = 'ENCODING_ERROR',
  INVALID_VPATH_PARENT_SEGMENT = 'INVALID_VPATH_PARENT_SEGMENT',
  INVALID_VPATH_FORMAT = 'INVALID_VPATH_FORMAT',
  STORE_CONFLICT = 'STORE_CONFLICT',
  UNKNOWN = 'UNKNOWN'
}

export enum IdentityPlatform {
  WINDOWS = 'WINDOWS',
  POSIX = 'POSIX',
  UNKNOWN = 'UNKNOWN'
}

export enum OsKind {
  WINDOWS = 'WINDOWS',
  POSIX = 'POSIX'
}

export enum CasePolicy {
  AUTO = 'AUTO',
  SENSITIVE = 'SENSITIVE',
  INSENSITIVE = 'INSENSITIVE'
}